+++ docs/changelog.md
@@ -2,6 +2,11 @@
 
 ## Upcoming release
 
+### For Content Pack modders
+
+- Added token for use objects from JsonAssets in quest triggers. 
+(example for ItemDelivery quest type: `Willy {{ja:Fish Oil}}`)
+
 ### For SMAPI modders
 
 - Added trigger parser-loader in CustomQuest class